Move some protocol definitions closer to their usage sites
